Move things to the lib folder
