+++ index.js
@@ -1,4 +1,4 @@
-const app = require('./server')()
+const app = require('./lib/server')()
 const port = process.env.PORT || 3000
 app.listen(port, () => {
   console.log('Listening at http://localhost:' + port)
