+++ lib/redis-bus.js
